Change SQL schemas
In this commit, both the schema.sql and data.sql are changed. The first
is changed to fix some syntactical errors within the file.

Data.sql is changed to not explicitly assign the IDs to the order. This
is done because these IDs existing orders are then not taken into
account when creating new orders, resulting in a duplicate PK violation
the first 7 tries of creating a new order.